Gemini/paixueji/requirements.txt
flask
flask-cors
google-genai
requests
pydantic
loguru